+++ ui/widgets/error-logs_templ.go
@@ -15,7 +15,7 @@ import (
 	"github.com/mjefferson-whs/listener/ui/components"
 )
 
-func ErrorLogs(totalCount int64, errorLogs []*data.Log) templ.Component {
+func ErrorLogs(totalCount int, errorLogs []*data.Log) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
